code/analysis/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
asn1crypto=0.24.0=py37_0
astroid=2.1.0=py37_0
autograd=1.2=pypi_0
blas=1.0=mkl
ca-certificates=2019.1.23=0
certifi=2018.11.29=py37_0
cffi=1.11.5=py37he75722e_1
chardet=3.0.4=py37_1
conda=4.6.7=py37_0
conda-env=2.6.0=1
contextlib2=0.5.5=pypi_0
cryptography=2.4.2=py37h1ba5d50_0
cycler=0.10.0=py37_0
dbus=1.13.6=h746ee38_0
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
future=0.17.1=pypi_0
glib=2.56.2=hd408876_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
icu=58.2=h9c2bf20_1
idna=2.8=py37_0
intel-openmp=2019.1=144
isort=4.3.8=py37_0
jpeg=9b=h024ee3a_2
kiwisolver=1.0.1=py37hf484d3e_0
lazy-object-proxy=1.3.1=py37h14c3975_2
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.36=hbc83047_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.10=h2733197_2
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=he19cac6_0
mag=0.1=pypi_0
matplotlib=3.0.2=py37h5429711_0
mccabe=0.6.1=py37_1
mkl=2019.1=144
mkl_fft=1.0.10=py37ha843d7b_0
mkl_random=1.0.2=py37hd81dba3_0
ncurses=6.1=he6710b0_1
ninja=1.8.2=py37h6bb024c_1
numpy=1.16.2=py37h7e9f1db_0
numpy-base=1.16.2=py37hde5b4d6_0
olefile=0.46=py37_0
openssl=1.1.1b=h7b6447c_0
opt-einsum=2.3.2=pypi_0
pcre=8.42=h439df22_0
pillow=5.4.1=py37h34e0f95_0
pip=18.1=py37_0
pycosat=0.6.3=py37h14c3975_0
pycparser=2.19=py37_0
pylint=2.2.2=py37_0
pyopenssl=18.0.0=py37_0
pyparsing=2.3.1=py37_0
pyqt=5.9.2=py37h05f1152_2
pyro-ppl=0.3.1.post1=pypi_0
pysocks=1.6.8=py37_0
python=3.7.1=h0371630_7
python-dateutil=2.8.0=py37_0
python-graphviz=0.10.1=pypi_0
pytorch-cpu=1.0.1=py3.7_cpu_2
pytz=2018.9=py37_0
pyzmq=18.0.1=pypi_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
requests=2.21.0=py37_0
ruamel_yaml=0.15.46=py37h14c3975_0
scipy=1.2.1=py37h7c811a0_0
setuptools=40.6.3=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.12.0=py37_0
sqlite=3.26.0=h7b6447c_0
tk=8.6.8=hbc83047_0
torchfile=0.1.0=pypi_0
torchvision-cpu=0.2.2=py_3
tornado=5.1.1=py37h7b6447c_0
tqdm=4.31.1=pypi_0
urllib3=1.24.1=py37_0
visdom=0.1.8.8=pypi_0
websocket-client=0.56.0=pypi_0
wheel=0.32.3=py37_0
wrapt=1.11.1=py37h7b6447c_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
